pyproject.toml
[project]
name = "validations"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.0",
    "numpy>=2.3.4",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "uvicorn>=0.38.0",
]
